refactor(bricks-search): replace grep_fast/glob_fast with glob_helpers + direct nexus._rust_compat
Phase F.0 of Phase 3 restructure plan.  The old wrappers
`bricks/search/primitives/grep_fast.py` and `glob_fast.py` carried two
distinct kinds of code:

  - thin Rust passthroughs (grep_bulk / grep_files_mmap /
    glob_match_bulk) that wrapped the kernel-resident lib-tier
    primitives with None-fallback branches that became dead once
    nexus_kernel is mandatory.
  - pure-Python glob query helpers (extract_static_prefix /
    is_simple_pattern / glob_match / glob_filter) used by the search
    tier for query optimization.

Splits them along that seam:

  - new src/nexus/bricks/search/primitives/glob_helpers.py owns the
    pure-Python helpers.  Same external API minus the dead Rust
    fallback branches (Rust path always taken).
  - the Rust passthroughs go away — callers import grep_bulk /
    grep_files_mmap / glob_match_bulk straight from nexus._rust_compat
    in subsequent commits.
  - rm grep_fast.py + glob_fast.py.
  - bricks/search/primitives/__init__.py re-exports glob_match (now
    sourced from glob_helpers) and grep_bulk (now sourced from
    nexus._rust_compat), keeping the public-facing
    `from nexus.bricks.search.primitives import grep_bulk` shape.

Sets up the SearchService / connector / factory / async_files /
benchmarks caller migration in the following commits.
