+++ src/nexus/bricks/search/primitives/__init__.py
@@ -1,23 +1,22 @@
 """Search primitives for Nexus.
 
-Low-level search utilities extracted from core/ to properly place them
-at brick tier rather than kernel tier (Issue #2123).
+Low-level search utilities at the brick tier (Issue #2123).
 
 These primitives power Search brick operations:
-- grep_fast: Fast content search with ripgrep
-- glob_fast: Fast file pattern matching
-- trigram_fast: Trigram-based search indexing
+- glob_helpers: pure-Python glob query helpers (extract_static_prefix,
+  is_simple_pattern, glob_match, glob_filter)
+- trigram_fast: Trigram-based search indexing (Rust-accelerated)
 
-Re-exported for convenience. All modules follow Rust-accelerated pattern
-with Python fallback when nexus_kernel is unavailable.
+The Rust-accelerated grep / glob primitives (`grep_bulk`,
+`grep_files_mmap`, `glob_match_bulk`) are imported directly from
+`nexus._rust_compat` — they are kernel-resident lib-tier helpers, not
+brick-tier wrappers.
 
 Related: NEXUS-LEGO-ARCHITECTURE.md (minimal kernel, maximal bricks)
 """
 
-from nexus.bricks.search.primitives.glob_fast import glob_match  # noqa: F401
-
-# Re-export primitives for public API
-from nexus.bricks.search.primitives.grep_fast import grep_bulk  # noqa: F401
+from nexus._rust_compat import grep_bulk  # noqa: F401
+from nexus.bricks.search.primitives.glob_helpers import glob_match  # noqa: F401
 from nexus.bricks.search.primitives.trigram_fast import (  # noqa: F401
     build_trigram_index,
     search_trigram,
@@ -27,9 +26,9 @@ from nexus.bricks.search.primitives.trigram_fast import (
 )
 
 __all__ = [
-    "grep_bulk",
-    "glob_match",
     "build_trigram_index",
+    "glob_match",
+    "grep_bulk",
     "search_trigram",
     "trigram_available",
 ]
